refactor(irm): remove oncallpublic client and oncalltypes sub-package
The IRM plugin proxy now accepts SA tokens (grafana/irm#9473), so the
separate public API client workaround is no longer needed. All OnCall
requests now go through the plugin proxy regardless of auth mode.

- delete oncallpublic/ (~1400 LOC) and oncalltypes/ sub-packages
- move types, constants, interface, and list options into irm package
- simplify LoadOnCallClient to always use the plugin proxy client
- remove dead oncall-url provider config key
- update architecture docs to reflect OnCall using k8s REST transport
