+++ internal/providers/irm/oncall_client_test.go
@@ -10,7 +10,6 @@ import (
 	"time"
 
 	"github.com/grafana/gcx/internal/providers/irm"
-	"github.com/grafana/gcx/internal/providers/irm/oncalltypes"
 )
 
 func newTestOnCallClient(t *testing.T, handler http.Handler) *irm.OnCallClient {
@@ -216,7 +215,7 @@ func TestListAlertGroups_StopsEarlyWithLimit(t *testing.T) {
 	srvURL = srv.URL
 
 	client := &irm.OnCallClient{HTTPClient: srv.Client(), Host: srv.URL}
-	items, err := client.ListAlertGroups(context.Background(), oncalltypes.WithLimit(1))
+	items, err := client.ListAlertGroups(context.Background(), irm.WithLimit(1))
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
@@ -247,7 +246,7 @@ func TestListAlertGroups_WithStartedAfter(t *testing.T) {
 	}))
 
 	cutoff := time.Date(2025, 1, 15, 0, 0, 0, 0, time.UTC)
-	items, err := client.ListAlertGroups(context.Background(), oncalltypes.WithStartedAfter(cutoff))
+	items, err := client.ListAlertGroups(context.Background(), irm.WithStartedAfter(cutoff))
 	if err != nil {
 		t.Fatalf("unexpected error: %v", err)
 	}
