--- internal/providers/irm/oncall_types_reexport.go
@@ -1,48 +1,0 @@
-package irm
-
-import "github.com/grafana/gcx/internal/providers/irm/oncalltypes"
-
-// Re-export types from oncalltypes so existing code in this package won't have
-// to change when we rip out the oncallpublic package and move all the
-// oncalltypes types back to this irm package.
-
-type (
-	OnCallAPI                 = oncalltypes.OnCallAPI
-	Integration               = oncalltypes.Integration
-	EscalationChain           = oncalltypes.EscalationChain
-	EscalationPolicy          = oncalltypes.EscalationPolicy
-	Schedule                  = oncalltypes.Schedule
-	Shift                     = oncalltypes.Shift
-	Route                     = oncalltypes.Route
-	Webhook                   = oncalltypes.Webhook
-	AlertGroup                = oncalltypes.AlertGroup
-	User                      = oncalltypes.User
-	Team                      = oncalltypes.Team
-	UserGroup                 = oncalltypes.UserGroup
-	SlackChannel              = oncalltypes.SlackChannel
-	Alert                     = oncalltypes.Alert
-	Organization              = oncalltypes.Organization
-	ResolutionNote            = oncalltypes.ResolutionNote
-	ShiftSwap                 = oncalltypes.ShiftSwap
-	FlatShift                 = oncalltypes.FlatShift
-	FinalShiftEvent           = oncalltypes.FinalShiftEvent
-	FilterEventsResponse      = oncalltypes.FilterEventsResponse
-	CreateResolutionNoteInput = oncalltypes.CreateResolutionNoteInput
-	UpdateResolutionNoteInput = oncalltypes.UpdateResolutionNoteInput
-	CreateShiftSwapInput      = oncalltypes.CreateShiftSwapInput
-	UpdateShiftSwapInput      = oncalltypes.UpdateShiftSwapInput
-	TakeShiftSwapInput        = oncalltypes.TakeShiftSwapInput
-	UserReference             = oncalltypes.UserReference
-	DirectPagingInput         = oncalltypes.DirectPagingInput
-	DirectPagingResult        = oncalltypes.DirectPagingResult
-	ShiftRequest              = oncalltypes.ShiftRequest
-)
-
-const (
-	APIGroup   = oncalltypes.APIGroup
-	APIVersion = oncalltypes.APIVersion
-	Version    = oncalltypes.Version
-)
-
-//nolint:gochecknoglobals // constant-like slice shared by commands and adapter registrations
-var DefaultStripFields = oncalltypes.DefaultStripFields
